backend/requirements.txt
asyncpg==0.30.0
eventlet==0.39.1
fastapi==0.115.12
psycopg2_binary==2.9.3
SQLAlchemy==2.0.40
uvicorn==0.33.0
tabulate==0.9.0
aiokafka==0.12.0
websockets==11.0.3
numpy==1.24.4
pandas==1.5.3
pytz==2022.7.1
requests==2.32.4
httpx==0.28.1

## pip install fastapi[all]